realtime-commentator/pyproject.toml
[project]
name = "realtime-commentator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs",
    "pydub",
    "python-dotenv",
]
